% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polyNDVI.R
\name{polyNDVI}
\alias{polyNDVI}
\title{Gather weekly NDVI for polygon areas}
\usage{
polyNDVI(yrs, wks, polyfolder)
}
\arguments{
\item{yrs}{vector c() or sequence seq() of years do be extracted}

\item{wks}{vector c() or sequence seq() of weeks do be extracted}

\item{polyfolder}{path to the folder where shape files of polygons are saved}
}
\value{
A object of the class SpatialPolygonsDataFrame
will be created. The extracted values are stored in the
slot data and the column
}
\description{
The Data from AVHRR is gathered from the FTP server
and NDVI values are extracted for polygon areas in
a SpatialPolygons object
}

