% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visNDVI.R
\name{visNDVI}
\alias{visNDVI}
\title{Visulize point coordinates with one NDVI map}
\usage{
visNDVI(yrs, wks, lonlim, latlim, lns = T)
}
\arguments{
\item{yrs}{vector c() or sequence seq() of years do be visualized}

\item{wks}{vector c() or sequence seq() of weeks do be visualized}

\item{lonlim}{longitude range c(min,max) for the plot}

\item{latlim}{latitude range c(min,max) for the plot}
}
\value{
A multi-panel graph of the NDVI rasters is
plotted.
}
\description{
Point coordinates are visualized with a weekly NDVI
map of choice.
}

